Add tests to verify autoproxy server object lifetime
An autoproxy consists of a server object, and a client
proxy. When the client proxy is released, the server
object should be as well.

Add a few simple tests to verify superficially
that this does indeed happen.

Note: these tests currently fail, so they are marked
xfail (expected to fail).